+++ src/cmos/cmosdump/cmosdump.c
@@ -129,7 +129,7 @@ static int cmosdump_test1(fwts_framework *fw)
 	unsigned char data[0x80];
 
 	/* Read CMOS Data */
-	for (i=0;i<sizeof(data); i++) {
+	for (i=0;i<(int)sizeof(data); i++) {
 		if (fwts_cmos_read(i, &data[i]) != FWTS_OK) {
 			fwts_log_error(fw, "Cannot get read/write permission on I/O ports.");
 			return FWTS_ERROR;
@@ -137,7 +137,7 @@ static int cmosdump_test1(fwts_framework *fw)
 	}
 
 	fwts_log_info_verbatum(fw, "CMOS Memory Dump:");
-	for (i=0;i<sizeof(data); i+= 8) {
+	for (i=0;i<(int)sizeof(data); i+= 8) {
 		char buffer[128];
 		fwts_dump_raw_data(buffer, sizeof(buffer), data + i, i, 16);
 		fwts_log_info_verbatum(fw, "%s", buffer);
